+++ src/test/java/com/fillumina/xjc/validation/MultiplePatternsWithBaseFixtureTest.java
@@ -13,6 +13,10 @@ class MultiplePatternsWithBaseFixtureTest extends FixtureTest {
         return Stream.of(
                 Case.of("MultiplePatternsWithBase", option("targetNamespace", "a")),
                 Case.of("MultiplePatternsWithBaseAsList", option("targetNamespace", "a"),
-                        option("patternList", "true")));
+                        option("patternList", "true")),
+                Case.of("MultiplePatternsWithBaseAsListWithOverride", option("targetNamespace", "a"),
+                        option("patternList", "true"),
+                        option("override", "*#singlePattern@Pattern:message = overridden"),
+                        option("override", "*#multiPatternInheritMultiBase@NotNull")));
     }
 }